rename data to dataDir
