--- packages/embodi/src/core/definitions/config.ts
@@ -1,11 +1,0 @@
-import type { Plugin as VitePlugin } from "vite";
-export interface EmbodiConfig {
-	statics: string;
-	base: string;
-	dist: string;
-	data: string;
-	source: `/${string}`;
-	templatePrefix: string;
-	publicDir: string;
-	plugins: VitePlugin[];
-}
